% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emp_mode}
\alias{emp_mode}
\title{Compute the univariate empirical mode via Bernstein polynomials.}
\usage{
emp_mode(data, threshold = NULL, smooth = TRUE,
  smooth_option = "optimalOfPdf", fix_lower = NULL, fix_upper = NULL,
  m_degree = NULL, m_knots = NULL)
}
\arguments{
\item{data}{A univariate sample data.}

\item{threshold}{A number of subsample size used in Anderson-Darling tests, and default is the size of the sample data.}

\item{smooth}{A indicator of if performing Bernstein polynomials smoothing or not; If TRUE (default), adapt Bernstein polynomials smoothing.}

\item{smooth_option}{A name of method of finding maximum based on the Bernstein polynomials density, and "optimalOfPdf" (default) indicates finding optimum on the smoothing density.}

\item{fix_lower}{A value of the lower bound of the the sample data; If NULL (default), set min{0, the minimum of the sample data}; Set a particualr value otherwise.}

\item{fix_upper}{A value of the upper bound of the the sample data; If NULL (default), set max{1, the maximum of the sample data}; Set a particualr value otherwise.}

\item{m_degree}{A number of degree of Bernstein polynomials; If NA (default), set the one with maximal p-value that less than 0.95; Set a particualr value otherwise.}

\item{m_knots}{A number of knots in finding the maximum of Bernstein polynomials density; If NA (default), set n / log(n); Set a particualr value otherwise.}
}
\value{
The empirical mode of \code{data}, the degree of Bernstein polynomials used and the corresponding p-value of the Anderson-Darling test based on the subsample and Bernstein polynomials.
}
\description{
Compute the univariate empirical mode via Bernstein polynomials.
}
\examples{
m_threshold = 1e3
n = 1e6
set.seed(1)
data = rbeta(n, 2, 5)
emp_mode(data, threshold = m_threshold)
}
